requirements.txt
matplotlib
seaborn
streamlit
streamlit-ace
openpyxl
openai
multipage-streamlit
pandas-gpt[openai]
streamlit-nested-layout